# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ERP RRI - API Documentation",
    "version": "1.0.0",
    "description": "REST API for ERP RRI system. Manages master data including barang, supplier, customer, COA, kontrak, jabatan, karyawan, and more. All endpoints require Bearer JWT authentication obtained from /login."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "500"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/inventory/stock-opname": {
      "get": {
        "operationId": "get-v1-inventory-stock-opname",
        "summary": "",
        "description": "",
        "tags": [
          "V1"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "operationId": "post-v1-inventory-stock-opname",
        "summary": "",
        "description": "",
        "tags": [
          "V1"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/v1/laporan/arus-kas/pdf": {
      "get": {
        "operationId": "get-v1-laporan-arus-kas-pdf",
        "summary": "",
        "description": "",
        "tags": [
          "V1"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/v1/laporan/laba-rugi/pdf": {
      "get": {
        "operationId": "get-v1-laporan-laba-rugi-pdf",
        "summary": "",
        "description": "",
        "tags": [
          "V1"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/v1/laporan/neraca/pdf": {
      "get": {
        "operationId": "get-v1-laporan-neraca-pdf",
        "summary": "",
        "description": "",
        "tags": [
          "V1"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
...
}